requirements.txt
beautifulsoup4
requests
python-telegram-bot==20.0a6
